Fix '_context' error in ProductController by adding dependency injection

+++ Repositories/EFProductRepository.cs
@@ -1,6 +1,8 @@
-﻿using Microsoft.EntityFrameworkCore;
-using PhanVuBaoMinh.Data;
+﻿using PhanVuBaoMinh.Data;
 using PhanVuBaoMinh.Models;
+using System.Collections.Generic;
+using System.Linq;
+using Microsoft.EntityFrameworkCore;
 
 namespace PhanVuBaoMinh.Repositories
 {
@@ -13,22 +15,37 @@ namespace PhanVuBaoMinh.Repositories
             _context = context;
         }
 
-        public async Task<IEnumerable<Product>> GetAllAsync()
+        public IEnumerable<Product> GetAll()
         {
-            return await _context.Products
-                       .Include(p => p.Category)
-                       .ToListAsync();
+            return _context.Products.Include(p => p.Category).ToList();
+        }
+
+        public Product GetById(int id)
+        {
+            return _context.Products.Include(p => p.Category).FirstOrDefault(p => p.Id == id);
         }
 
         public async Task<Product> GetByIdAsync(int id)
         {
-            return await _context.Products.FindAsync(id);
+            return await _context.Products.Include(p => p.Category).FirstOrDefaultAsync(p => p.Id == id);
+        }
+
+        public void Add(Product product)
+        {
+            _context.Products.Add(product);
+            _context.SaveChanges();
         }
 
         public async Task AddAsync(Product product)
         {
-            _context.Products.Add(product);
+            await _context.Products.AddAsync(product);
             await _context.SaveChangesAsync();
+        }
+
+        public void Update(Product product)
+        {
+            _context.Products.Update(product);
+            _context.SaveChanges();
         }
 
         public async Task UpdateAsync(Product product)
@@ -37,18 +54,24 @@ namespace PhanVuBaoMinh.Repositories
             await _context.SaveChangesAsync();
         }
 
-        // Phương thức Update đồng bộ
-        public void Update(Product product)
+        public void Delete(int id)
         {
-            _context.Products.Update(product);
-            _context.SaveChanges();
+            var product = _context.Products.FirstOrDefault(p => p.Id == id);
+            if (product != null)
+            {
+                _context.Products.Remove(product);
+                _context.SaveChanges();
+            }
         }
 
         public async Task DeleteAsync(int id)
         {
             var product = await _context.Products.FindAsync(id);
-            _context.Products.Remove(product);
-            await _context.SaveChangesAsync();
+            if (product != null)
+            {
+                _context.Products.Remove(product);
+                await _context.SaveChangesAsync();
+            }
         }
     }
 }
